make 'glCompSet.textureCount' a 'size_t'
Squashes 2 -Wsign-conversion warnings.

+++ lib/glcomp/glcompdefs.h
@@ -305,7 +305,7 @@ extern "C" {
 
 	glCompObj **obj;
 	size_t objcnt;
-	int textureCount;
+	size_t textureCount;
 	glCompTex **textures;
 	glCompMouse mouse;
     } glCompSet;
